+++ src/__playground__/worker.ts
@@ -5,4 +5,5 @@ export const api = await define({
     seek: () => {
         console.log(workerData);
     },
+    add5: (num: number) => num + 5,
 });